widen line special args from byte to int
Hexen's binary format stores line args as bytes, and the engine
mirrored that in line_t and the whole executor chain (byte *args
through every P_Execute*/EV_* signature).  ZDoom UDMF removes the
limit: args are full ints, and real maps rely on it -- myhouse.pk3's
MAP01 carries ~1900 linedefs whose args exceed 255 (line ids, sector
tags and TIDs in a 143k-line map, negative polyobj angles), so byte
storage truncates door tags and teleport destinations.

Widen line_t.args to int and sweep the parameter chain (map_format
dispatch, hexen specials/po_man/ACS, heretic A_LocalQuake) from
byte* to int*.  The on-disk doomdata.h structs stay byte -- they
mirror the wad format -- and every load path already copies args
element-wise.  mobj special_args and the hexen thing staging were
already int.  Hexen savegames wrote args via SV_WriteLong and only
narrowed on read; drop that (byte) cast so saved args round-trip at
full width.

No behavior change for byte-ranged values: Hexen, Heretic and Doom
gameplay probes pass identically.

+++ src/heretic/p_action.c
@@ -1976,7 +1976,7 @@ void A_Quake(mobj_t * actor)
     }
 }
 
-dbool A_LocalQuake(byte * args, mobj_t * actor)
+dbool A_LocalQuake(int * args, mobj_t * actor)
 {
     mobj_t *focus, *target;
     int     lastfound = -1;
